Remove the definitions of ready and done
Remove the definitions of ready and done as they were incomplete and
unused. Most teams have their own definitions of ready and done on
Confluence. Furthermore, as these are not engineering standards, but
rather team process documentation relating to non-engineering roles on
the team they should live on Confluence anyway.